dash/src/requirements.txt
dash
dash_core_components
dash_html_components
pandas
psycopg2-binary
plotly